% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_colors.R
\name{fetch_color}
\alias{fetch_color}
\title{Fetch colors}
\usage{
fetch_color(
  n = 1,
  type = c(names(color_list), "random"),
  tag = names(color_list[[type]]),
  is.extend = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{n}{Number of colors (>= 1)}

\item{type}{Which plot type you want?
Use \code{names(color_list)} to check built-in plot types.}

\item{tag}{Which preset color palette to use?}

\item{is.extend}{Extend the palette to n.}

\item{verbose}{Show the verbose}
}
\description{
Get colors from preset color list or \code{\link{randomcoloR}} package.
}
