% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_colors.R
\name{fetch_random_color}
\alias{fetch_random_color}
\title{Get random colors}
\usage{
fetch_random_color(n = 1, usepalette = FALSE, ...)
}
\arguments{
\item{n}{Number of colors (>= 1)}

\item{usepalette}{Whether or not to use \code{\link{distinctColorPalette}}}

\item{...}{Parameters passed to \code{\link{randomColor}}.}
}
\description{
Get random colors from \code{\link{randomcoloR}} package.
}
