% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_list.R
\docType{data}
\name{color_list}
\alias{color_list}
\title{The built-in list of color sets}
\format{
A list with the following preset palettes
\describe{
  \item{venn}
  \item{line}
  \item{pie}
  \item{twopie}
  \item{bar}
  \item{hist}
  \item{heatmap}
  \item{dot}
  \item{box}
  \item{all}
  \item{gray}
  \item{onlinereport}
  \item{tsne}
  \item{ggsci}
}
}
\usage{
color_list
}
\description{
The built-in list of color sets used for different plot types.
}
\keyword{datasets}
